Refactor combat logic and restructure files
Extract combat-related logic into a new `combat` module, simplifying `gameLogic`. Rename `gameData` to `events` for clarity, and move `generatePrices` back to `gameLogic`. Define `GameState` type in `GameContext` to centralize state management. This refactor improves code organization and maintainability by separating concerns and clarifying module purposes.
